Allow retrying of multi-po-downloads and make single-downloads more reliable
If the checkout of a SVN dir failed, instead of only allowing ignoring
or canceling the entire process, we now allow to retry the checkout
any number of time.

For single file exports we additionally try to do a better job at guessing
whether file exists or not. If the export fails we run an info command
and only if that fails too we assume that the file does not exist.
If the file exists at first try of info we let the user retry,
if it does not exist at first info test we simply assume the file does
not exist and move on.