implement a simply vcs::clean! and source::clean ability
rationale being that we want to get rid of .svn and .git in an opaque
manner.
source::clean is not exactly what it should be and how it should be used..
ultimately you'd order sources in a list (project source < l10n source <
doc source < whatever source) and then have a release/project run
.get on all sources and then .clean on all sources and then archive the
target directory followed by .cleanup on all sources (in reverse order
obviously).

the other approach would be nesting sources which is equally compelling
but possibly makes for sillier code and more ways to make things go wrong