strip fuzzy obsoletion as well as regular obsoletion
http://mail.kde.org/pipermail/plasma-devel/2014-July/033535.html

gettext strings are marked obsolete by starting the line with #~, but
sometimes an obsoletion is preceded by a fuzzy line which previously would
have left the fuzzy mark behind when removing obsoletion. so, the
processing was a bit improved to first remove the fuzzies and then all
remaining obsolete lines.

this should prevent garbage being left behind