l10n: thread .get to increase speed substantially
svn is super slow in establishing the connection. just doing the connection
takes usually more than *one* second on my system!
considering we do a but load of atomic queries as part of the get process
we already get substantial speed improvements by simply threading all of
this as it makes the connection setup have a much smaller impact.

- thread count is limited by anonsvn's connection limit
- there are various non-thread-safe variable uses that don't matter
  for now as ruby2 still has a GIL
- also needs implementation in documentation