work around inconsistency in l10n mapping of projects
kde/workspace -> kde-workspace
kde/kdenetwork -> kdenetwork
extragear/kitten -> extragear-kitten

as a result if the first part is kde and the second part is not workspace
we drop the kde part (e.g. kdenetwork rather than kde-kdenetwork). if the
first part is not kde we merge without dropping (e.g. playground- and
extragear-)