change the way compat works to be less ad-hoc
mapping as classes get defined has a major problem due to how requires work

proc.rb
  require 'releaseme/x'
    require './z'
  require 'y'
    require 'releaseme/z'

In this line up we'd initially load X which loads Z at this point both X
and Z are defined. We later load Y via compat files, this loads Z, BUT
Z is already loaded so there won't be no events as the class already exists
and thus we cannot setup compat mapping for Z because we don't even know
Z is a thing

To solve this we'll actually make the compat mapping code more readable
and reliable by not doing an event listener but simply require the actual
file and then iterate all consts in the ReleaseMe module. The consts
we haven't yet define in the root scope we'll map the others we'll ignore.

This means that no matter what you do whenever a compat require is used
it will absolutely make sure everything that can be defined is defined.