try to break compat nonesense when gemspecing through bundler
there is no end to the madness!

bundler is running the gemspec but doesn't actually expect anything of it
and ignores the files list. it basically dumps the cloned tree into some
random path it will load it from. that technically would be fine BUT that
means `require 'git'` may in fact load the compat git.

so, instead try to mangle the dir at hand to reject files iff the gemspec
is invoked by bundler