build new script downloader to deal with transcript files
also try something new design-wise, the l10n class always felt like a mess
which may be better if we isolate the super-rigging of "fetching all l10n"
from the individual implementations of how that is achieved.

notably we'd have downloads that get pre-created and then call download
on all of them expecting to dump themself into a tmpdir

in this set up we'd have
SinglePotDownloader - gets one file but only if one template
MultiPotDownloader - gets multiple files but only if many templtaes
ScriptDownloader - gets scripts regardless of templte count

additionally the thread handling needs spreading further, which is tricky
because technically none of the things we do are in fact thread-safe and
so we have these sweepingly huge thread blocks because technically we cant
break things up. incidentally that also would be easier to do without
looking like crap with more classes on the table.

anyway, something for the future