build caching logic into scripts fetching
this is fairly expensive as we need to iter languages*templates so use
a simple list cache to avoid needless svn calls.

the l10n class creates a cache which iters languages once to get all
templates in the scripts dir (per-language). the scripts downloader later
simply checks the cache if a given template is available for a language,
if not that's the end of that