detach currently-building from quality check
the motivation of the quality check is to enable accepting subpar results
as such we should also ignore whether or not a job is still building.

e.g. in plasma when the level is lowered it should only print a warning
iff the job's latest *completed* build is red, all other cases are
uninteresting at this point