Implement qt move. we let poqm/ become po/ to align with frameworks
in framworks there's sometimes sources that only use qt translations, they
use po/ all the same. so instead of forcing the application double-dir
approach and possibly breaking something align the behavior.

as far as applications are concerned either will work fine as the release
tools mangle the cmakelists, in case of framworks the po/ stuff is properly
controled by the developers in git so no mangling occurs