reduce drop limit after removing a leading kde/ part
The general assumption is that path.split.pop.join('-') will be the
i18n_path. e.g. playground/utils becomes playground-utils and
playground/utils/libs becomes also playground-utils. To that end we drop
parts until we have <= 2 to join. This was going wrong with nested projects
inside kde/* projects as we drop the first kde/ part.
So kde/kdegraphics/libs would be [kde,kdegraphics,libs] then we drop the
kde to get [kdegraphics,libs] and now no longer drop parts treating it
like playground, when really it shouldn't be as effectively our
requirements are different. Technically we want [kde,kdegraphics,libs] to
be popped until only 2 are left, we can't do that because dropping kde
is sort of a pre condition, otherwise we'll have to do some mind bending
to prevent empty paths etc.

BUG: 379161