source no longer expects that multiple clones pass
this expectation puzzles me a bit. it was part of the initial import but
has no clear purpose. today it certainly makes no sense because taring
will first clean up the target directory if it already exists.
from an outside POV a pre-existing dir is thrown away and then cloned
anew (so as to be pristine).

drop this expectation in light of not understanding its purpose and it
no longer being true since previous commit that made failed clones actually
fatal. should we re-discover the use we can always revisit this