print l10nstatistics and use stats to drop 0-coverage l10n
translators may have copies of templates but not have translated them, so
they have effectively 0 coverage (e.g. currently in phonon trunk). this
means we ship translations which contain nothing and leading to wasted
time compiling them and also every time loading them and trying to look
up strings in them. seeing as they add no value we can simply drop them...

the removal entirely anchors on strings... iff a language had any *.po
file and all available .po files amount to 0 coverage the language as a
whole is considered incomplete. this also extends to scripts and data
assets because the assumption is that a language that doesn't have any
translated strings will offer poor quality regardless of the quality
of additional data assets anyway.
for the time being a language which only has data assets is ignored here
because we cannot qualify them, so we assume they are complete and useful.

this leaves overall statistics with a single remaining shortcoming:
lack of understanding of actual translated ness. translated percentage is
calculated from the available strings **in that language**. IOW if a
language only translates 1 of 10 templates it should be ~incomplete but
currently if that 1 language is complete we think the language as a whole
is. to fix this properly there's some additional magic needed to determine
the overall available strings (also see comment in l10n.rb)