throw away nesting support
this was originally intended to allow batch releasing of modules but
that turned out to be useless since projects can move between modules
with legacy releases being in one module and preview releases being
in another rendering this a useless feature

nobody seems to rely on it from a quick thread on the release team list

from_xpath is now deprecated in favor of from_find. passing by absolute
path is no longer supported and as a side effect all projects are now
presumed implicitly unique. should there be projects for which that is not
the case we'll raise in from_find now

BUG: 420501