throw away legacy xml support
we've been using the api exclusively for years and the xml isn't meant
to be used anyway, so throw away the code and flatten out the project.rb
again by merging the API code in (was isolated in separate files before)

project no longer carries project_element as that only made sense for xml.
the helper class projectsfile is no longer needed and also was removed.