convert the entire l10n test to mimic the new svn structure
this moves away from the legacy structure.

while the code was dynamic enough to deal with both the new and the old
without requiring changes, aligning the expectations of the test with
the reality in production will make it easier moving forward to add new
tests.
also makes the structure more expected

this also changes the i18n_paths to be kinda the same as the project
names, as-is the case in svn. the paths are effectively the project id
as defined in the repo metadata. that happens to usually be the name but
may not be as names are not unique outside their nested structure on
invent

e.g. mobile/phone, maui/phone would be i18n path mobile-phone and
maui-phone rather than phone as that'd not be unique