move to filter checkouts instead of checking out selectively
with the new l10n structure we can simply checkout entire directories
instead of trying to detect the templates used by the source and then
selectivly getting the ones we need from svn to avoid having to checkout
entire modules where 99% of the templates won't even be appicable.
since each repo now has their own directory we can checkout that dir and
call it a day. simplifies the code a bit and also means template
detection is no longer necessary

BUG: 424031