remove all pot (template) code
with the last of the users of dynamic template handling gone we can
throw away the entire template detection logic.
in a modern world the svn structure sports a dedicated directory per git
repository and inside those dirs are only useful things for that git
repo. that simplifies releasing as we can now much more liberally
checkout entire paths without fear of significant time getting wasted on
unnecessary data

(there is still some waste because we download files we'll later filter
out such as _desktop_ file translations, but it's not worth holding on
to all the extra code template code just for that)