fix ruby 3 compat
there may or may not be a bug with process status duplication preventing
the newly created instance from initializing completely. could well be
intentional though, copying status is a tad silly anyway as a status
object **ought** to be immutable after process execution.

the dup call came from a reentrant fix where previously `@status` was
captured directly and supposedly that caused trouble with the multi
threading of l10n as later in the run() functions we'd refer to `status`
which was then going through the accessor but that may already have been
modified by another thread. so the important change there was capturing
into a local variable `status` instead of `@status` directly.

to be on the safe side let's freeze the status object even though
reentrancy-wise the local variable is the important bit and statuses
should be always frozen anyway