add compat mapping code for phonon
for incredibly legacy reasons the tarball has a different name than the
repo which is incredibly annoying for me because I always have to hack
in some temporary compatibility. instead add a persistent compat map
that rejiggers the names accordingly. this also has a huge warning about
not abusing this feature for anything else. the entire release team is
pretty much in agreement over tar names always being the name of the
repo

ideally phonon should get its act together, but changing the repo name
will annoy translators while changing the tar name will annoy distros.
considering pretty much only phonon appears plagued by this it's not
worth the hassle (for now - that is: I'd like the shorter repo names as
tarball names, but changing names in a stable release update is very
awkward for distros)