remove the compatibility library
it's been fatally failing for the past couple of weeks and warning for
literal years. if software hasn't been ported yet it can break for all I
care.

fixing breakage related to this is merely a matter of either requiring
'lib/releaseme' instead of 'lib/whateverrandomfile' to simply use
everything (which is likely the most reasonable thing to do) or using
'lib/releaseme/whateverrandomfile' to continue using that one specific
file